pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "finnhub-python>=2.4.23",
    "newsapi-python>=0.2.7",
    "openai>=1.66.3",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "streamlit>=1.43.2",
    "stripe>=11.6.0",
    "trafilatura>=2.0.0",
    "yfinance>=0.2.54",
]
